fixed copy failure; moved logic into copy.lua
